Added support for autolinking and removed older versions
