+++ package.json
@@ -8,7 +8,7 @@
   },
   "peerDependencies": {
     "react-native": ">=0.56",
-    "react-native-windows": ">=0.61.5"
+    "react-native-windows": ">=0.64.0"
   },
   "keywords": [
     "Crypto.getRandomValues",
