--- windows/RNGetRandomValues/ReactPackageProvider.cpp
@@ -1,17 +1,0 @@
-#include "pch.h"
-#include "ReactPackageProvider.h"
-#include "ReactPackageProvider.g.cpp"
-
-#include <ModuleRegistration.h>
-
-// NOTE: You must include the headers of your native modules here in
-// order for the AddAttributedModules call below to find them.
-#include "RNGetRandomValues.h"
-
-namespace winrt::RNGetRandomValues::implementation
-{
-    void ReactPackageProvider::CreatePackage(IReactPackageBuilder const& packageBuilder) noexcept
-    {
-        AddAttributedModules(packageBuilder);
-    }
-}
