--- windows/RNGetRandomValues/ReactPackageProvider.h
@@ -1,20 +1,0 @@
-#pragma once
-
-#include "ReactPackageProvider.g.h"
-
-using namespace winrt::Microsoft::ReactNative;
-
-namespace winrt::RNGetRandomValues::implementation
-{
-    struct ReactPackageProvider : ReactPackageProviderT<ReactPackageProvider>
-    {
-        ReactPackageProvider() = default;
-
-        void CreatePackage(IReactPackageBuilder const& packageBuilder) noexcept;
-    };
-}
-
-namespace winrt::RNGetRandomValues::factory_implementation
-{
-    struct ReactPackageProvider : ReactPackageProviderT<ReactPackageProvider, implementation::ReactPackageProvider> {};
-}
