--- windows/RNGetRandomValues/ReactPackageProvider.idl
@@ -1,9 +1,0 @@
-namespace RNGetRandomValues
-{
-[webhosthidden]
-[default_interface]
-runtimeclass ReactPackageProvider : Microsoft.ReactNative.IReactPackageProvider
-{
-    ReactPackageProvider();
-};
-}
