--- windows/RNGetRandomValues61/RNGetRandomValues61.vcxproj.filters
@@ -1,33 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <ItemGroup>
-    <Filter Include="Resources">
-      <UniqueIdentifier>accd3aa8-1ba0-4223-9bbe-0c431709210b</UniqueIdentifier>
-      <Extensions>rc;ico;cur;bmp;dlg;rc2;rct;bin;rgs;gif;jpg;jpeg;jpe;resx;tga;tiff;tif;png;wav;mfcribbon-ms</Extensions>
-    </Filter>
-    <Filter Include="Generated Files">
-      <UniqueIdentifier>{926ab91d-31b4-48c3-b9a4-e681349f27f0}</UniqueIdentifier>
-    </Filter>
-  </ItemGroup>
-  <ItemGroup>
-    <ClCompile Include="..\RNGetRandomValues\pch.cpp" />
-    <ClCompile Include="$(GeneratedFilesDir)module.g.cpp" />
-    <ClCompile Include="..\RNGetRandomValues\RNGetRandomValues.cpp" />
-    <ClCompile Include="..\RNGetRandomValues\ReactPackageProvider.cpp" />
-  </ItemGroup>
-  <ItemGroup>
-    <ClInclude Include="..\RNGetRandomValues\pch.h" />
-    <ClInclude Include="..\RNGetRandomValues\ReactPackageProvider.h" />
-    <ClInclude Include="..\RNGetRandomValues\RNGetRandomValues.h" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="..\RNGetRandomValues\RNGetRandomValues.def" />
-    <None Include="packages.config" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="PropertySheet.props" />
-  </ItemGroup>
-  <ItemGroup>
-    <Midl Include="..\RNGetRandomValues\ReactPackageProvider.idl" />
-  </ItemGroup>
-</Project>
