+++ windows/ReactNativeGetRandomValues/ReactNativeGetRandomValues.h
@@ -1,0 +1,13 @@
+#pragma once
+#include "pch.h"
+#include "NativeModules.h"
+
+namespace winrt::ReactNativeGetRandomValues
+{
+    REACT_MODULE(ReactNativeGetRandomValues, L"RNGetRandomValues");
+    struct ReactNativeGetRandomValues
+    {
+        REACT_METHOD(getRandomBase64)
+        std::string getRandomBase64(int byteLength) noexcept;
+    };
+}
