+++ windows/ReactNativeGetRandomValues/ReactPackageProvider.h
@@ -1,0 +1,21 @@
+#pragma once
+#include "ReactPackageProvider.g.h"
+
+using namespace winrt::Microsoft::ReactNative;
+
+namespace winrt::ReactNativeGetRandomValues::implementation
+{
+    struct ReactPackageProvider : ReactPackageProviderT<ReactPackageProvider>
+    {
+        ReactPackageProvider() = default;
+
+        void CreatePackage(IReactPackageBuilder const &packageBuilder) noexcept;
+    };
+} // namespace winrt::ReactNativeGetRandomValues::implementation
+
+namespace winrt::ReactNativeGetRandomValues::factory_implementation
+{
+
+struct ReactPackageProvider : ReactPackageProviderT<ReactPackageProvider, implementation::ReactPackageProvider> {};
+
+} // namespace winrt::ReactNativeGetRandomValues::factory_implementation
