+++ windows/ReactNativeGetRandomValues/ReactPackageProvider.idl
@@ -1,0 +1,9 @@
+namespace ReactNativeGetRandomValues
+{
+    [webhosthidden]
+    [default_interface]
+    runtimeclass ReactPackageProvider : Microsoft.ReactNative.IReactPackageProvider
+    {
+        ReactPackageProvider();
+    };
+}
