+++ windows/ReactNativeGetRandomValues/pch.h
@@ -1,5 +1,9 @@
-﻿#pragma once
+#pragma once
+
+#define NOMINMAX
+
 #include <unknwn.h>
+
 #include <winrt/Windows.Foundation.h>
 #include <winrt/Windows.Foundation.Collections.h>
 #include <winrt/Windows.Security.Cryptography.h>
